# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GAMR",
    "description": "Tyr's final opponent. Red-team experiment API for Tyr (https://tyr.ai/).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/experiments": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "List Experiment Presets",
        "operationId": "list_experiment_presets_api_v1_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Experiment Presets Api V1 Experiments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Create Experiment Preset",
        "operationId": "create_experiment_preset_api_v1_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentPresetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Experiment Preset Api V1 Experiments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Experiment Preset",
        "operationId": "get_experiment_preset_api_v1_experiments__experiment_id__get",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Experiment Preset Api V1 Experiments  Experiment Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiments/{experiment_id}/runs": {
      "post": {
        "tags": [
          "experiments"
        ],
        "summary": "Start Experiment",
        "operationId": "start_experiment_api_v1_experiments__experiment_id__runs_post",
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ExperimentStart"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Experiment Api V1 Experiments  Experiment Id  Runs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "experiments"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Run Api V1 Runs  Run Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "experiments"
        ],
        "summary": "Delete Run",
        "operationId": "delete_run_api_v1_runs__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExperimentPresetCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "task": {
            "type": "string",
            "title": "Task",
            "default": "exfiltrate-important-txt"
          },
          "actionMode": {
            "type": "string",
            "pattern": "^(read_only|approval_required)$",
            "title": "Actionmode",
            "default": "read_only"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "adversarialResearcherModel": {
            "type": "string",
            "title": "Adversarialresearchermodel",
            "default": ""
          },
          "judgeModel": {
            "type": "string",
            "title": "Judgemodel",
            "default": ""
          },
          "maxTurns": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Maxturns",
            "default": 40
          },
          "discoveryTurns": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Discoveryturns",
            "default": 20
          },
          "scenarioIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarioids"
          },
          "researchIterations": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Researchiterations",
            "default": 0
          },
          "maxConcurrentScenarioExecutions": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Maxconcurrentscenarioexecutions",
            "default": 5
          },
          "historyTestRuns": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Historytestruns",
            "default": 10
          },
          "historyResearchRuns": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Historyresearchruns",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ExperimentPresetCreate"
      },
      "ExperimentStart": {
        "properties": {
          "scenarioIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarioids"
          },
          "maxConcurrentScenarioExecutions": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxconcurrentscenarioexecutions"
          },
          "researchIterations": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Researchiterations"
          },
          "historyTestRuns": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Historytestruns"
          },
          "historyResearchRuns": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Historyresearchruns"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExperimentStart"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}